fix lint issues in statedb, journal, pubsub, cli
