+++ x/evm/client/cli/utils.go
@@ -45,20 +45,3 @@ func formatKeyToHash(key string) string {
 
 	return ethkey.Hex()
 }
-
-// nolint: deadcode
-func cosmosAddressFromArg(addr string) (sdk.AccAddress, error) {
-	if strings.HasPrefix(addr, sdk.GetConfig().GetBech32AccountAddrPrefix()) {
-		// Check to see if address is Cosmos bech32 formatted
-		toAddr, err := sdk.AccAddressFromBech32(addr)
-		if err != nil {
-			return nil, errors.Wrap(err, "invalid bech32 formatted address")
-		}
-		return toAddr, nil
-	}
-
-	// Strip 0x prefix if exists
-	addr = strings.TrimPrefix(addr, "0x")
-
-	return sdk.AccAddressFromHex(addr)
-}
