+++ x/evm/keeper/statedb.go
@@ -340,7 +340,7 @@ func (k *Keeper) GetCommittedState(addr common.Address, hash common.Hash) common
 	return common.BytesToHash(value)
 }
 
-// GetState returns the committed state for the given key hash, as all changes are commited directly
+// GetState returns the committed state for the given key hash, as all changes are committed directly
 // to the KVStore.
 func (k *Keeper) GetState(addr common.Address, hash common.Hash) common.Hash {
 	return k.GetCommittedState(addr, hash)
